+++ NEWS.md
@@ -1,10 +1,12 @@
-CopernicusMarine v0.4.4.0010
+CopernicusMarine v0.4.4.0011
 -------------
 
  * Added `vignette("product-info")`
- * Fix for [issue 143](https://github.com/pepijn-devries/CopernicusMarine/issues/143)
+ * Omitting subset-arguments in `cms_download_subset()` calls is now allowed
+ * Removed deprecated argument 'crop' in `cms_download_subset()`
  * Added warnings for mismatch between requested and available
    dimension ranges for `cms_download_subset()`
+ * Fix for [issue 143](https://github.com/pepijn-devries/CopernicusMarine/issues/143)
  * Added tests
  * Updated documentation